requirements.txt
accelerate==0.24.0
bert-score==0.3.13
botocore==1.31.73
git+https://github.com/openai/CLIP.git
cloudpathlib==0.16.0
cmake==  3.25.0
cycler== 0.12.1
cymem==  2.0.8
datasets==  2.4.0
decorator== 5.1.1
decord== 0.6.0
deepspeed== 0.11.1
Deprecated==1.2.14
diffusers== 0.16.0
dill==   0.3.5.1
distlib==0.3.7
docopt== 0.6.2
eagerpy==0.30.0
editdistance==0.6.2
einops== 0.7.0
fonttools== 4.43.1
frozenlist==1.4.0
fsspec== 2024.2.0
ftfy==   6.1.1
gast==   0.5.4
gdown==  4.7.1
gensim== 4.3.2
gitdb==  4.0.11
GitPython== 3.1.40
google-api-core==2.12.0
google-auth== 2.23.3
google-auth-oauthlib==0.4.6
google-cloud-aiplatform==1.36.0
google-cloud-bigquery==  3.13.0
google-cloud-core==   2.3.3
google-cloud-domains==1.5.3
google-cloud-resource-manager==1.10.4
google-cloud-speech== 2.22.0
google-cloud-storage==2.13.0
google-cloud-vision== 3.4.5
google-crc32c==  1.5.0
google-pasta==0.2.0
google-resumable-media== 2.6.0
googleapis-common-protos==1.61.0
gradio_client==0.7.0
grpc-google-iam-v1==0.12.6
grpcio==1.59.2
grpcio-status==1.59.2
h5py==3.10.0
hjson==3.1.0
httpcore==1.0.2
httpx==0.25.1
huggingface-hub==0.17.3
identify==2.5.31
idna==3.4
imageio==2.31.6
importlib-metadata==6.8.0
ipython==8.17.2
Jinja2== 3.1.2
jmespath==  1.0.1
joblib== 1.3.2
kaggle== 1.5.16
keras==  2.8.0
Keras-Preprocessing== 1.1.2
kiwisolver==1.4.5
langcodes== 3.3.0
langdetect==1.0.9
language-tool-python==2.7.1
lazy_loader== 0.3
lemminflect== 0.2.3
libclang==  16.0.6
lit==    15.0.7
lru-dict==  1.2.0
lxml==   4.9.3
Markdown==  3.5
markdown-it-py== 3.0.0
MarkupSafe==2.1.2
matplotlib==3.8.0
matplotlib-inline==   0.1.6
mdurl==  0.1.2
more-itertools== 10.1.0
mpi4py== 3.1.4
mpld3==  0.5.9
mpmath== 1.3.0
multidict== 6.0.4
multiprocess==0.70.13
murmurhash==1.0.10
networkx==  3.0
ninja==  1.11.1.1
nltk==   3.8.1
nodeenv==1.8.0
num2words== 0.5.13
numpy==  1.24.1
oauthlib==  3.2.2
omegaconf== 2.3.0
opencv-python==  4.8.1.78
opencv-python-headless== 4.5.5.64
opendatasets==0.1.22
OpenHowNet==2.0
opt-einsum==3.3.0
packaging== 23.2
pandas== 2.1.2
parso==  0.8.3
peft==   0.6.1
pexpect==4.8.0
Pillow== 9.3.0
pinyin== 0.4.0
pip==    23.3
platformdirs==3.11.0
plotly== 5.18.0
portalocker== 2.8.2
pptree== 3.1
pre-commit==3.5.0
preshed==3.0.9
prompt-toolkit== 3.0.40
proto-plus==1.22.3
protobuf==  3.20.2
psutil== 5.9.6
ptyprocess==0.7.0
pure-eval== 0.2.2
py-cpuinfo==9.0.0
pyarrow==13.0.0
pyasn1== 0.5.0
pyasn1-modules== 0.3.0
pycld2== 0.41
pycocoevalcap==  1.2
pycocotools== 2.0.7
pydantic==  1.10.13
pydeck== 0.8.1b0
Pygments==  2.16.1
pyparsing== 3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-magic==0.4.27
python-slugify== 8.0.1
pytorch_revgrad==0.2.0
pytz==   2023.3.post1
PyYAML== 6.0.1
referencing== 0.30.2
regex==  2023.10.3
requests==  2.28.1
requests-oauthlib==   1.3.1
responses== 0.18.0
rich==   13.6.0
rpds-py==0.12.0
rsa==    4.9
s3transfer==0.7.0
safetensors== 0.4.0
salesforce-lavis==    1.0.1
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==  1.11.3
seaborn==0.13.0
segtok== 1.5.11
semver== 3.0.2
sentencepiece==  0.1.99
setuptools==68.0.0
shapely==2.0.2
six==    1.16.0
smart-open==6.4.0
smmap==  5.0.1
sniffio==1.3.0
soupsieve== 2.5
spacy==  3.7.2
spacy-legacy==3.0.12
spacy-loggers==  1.0.5
sqlitedict==2.1.0
srsly==  2.4.8
stack-data==0.6.3
streamlit== 1.28.1
SwissArmyTransformer==0.4.8
sympy==  1.12
tabulate==  0.9.0
tenacity==  8.2.3
tensorboard== 2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit== 1.8.1
tensorboardX==2.6.2.2
tensorflow==2.8.1
tensorflow-estimator==2.8.0
tensorflow-hub== 0.15.0
tensorflow-io-gcs-filesystem       0.34.0
tensorflow-text==2.8.1
termcolor== 2.3.0
terminaltables== 3.1.10
text-unidecode== 1.3
textattack==0.3.8
thinc==  8.2.1
threadpoolctl==  3.2.0
tifffile==2023.9.26
tiktoken==0.5.1
timm==0.4.12
tokenizers==0.14.1
toml== 0.10.2
tomli== 2.0.1
toolz== 0.12.0
torch== 2.0.1+cu118
torch-dct== 0.1.6
torchaudio==2.0.2+cu118
torchvision== 0.15.2+cu118
tornado==6.3.3
tqdm== 4.66.1
traitlets== 5.13.0
transformer-smaller-training-vocab==0.3.2
transformers==4.34.1
transformers-stream-generator==0.0.4
triton== 2.0.0
typer==  0.9.0
typing_extensions==   4.4.0
tzdata== 2023.3
tzlocal==5.2
urllib3==1.26.13
validators==0.22.0
virtualenv==20.24.6
Wand==   0.6.11
wasabi== 1.1.2
watchdog==  3.0.0
wcwidth==0.2.8
weasel== 0.3.3
webdataset==0.2.68
webencodings==0.5.1
websockets==11.0.3
Werkzeug==  3.0.1
wheel==  0.41.2
Wikipedia-API==  0.6.0
word2number== 1.1
wrapt==  1.15.0
xformers==  0.0.22

